refactor(app): refactor index.tsx, app new component in project
