+++ src/App.tsx
@@ -1,0 +1,50 @@
+import clsx from 'clsx';
+import { CSSProperties, useState } from 'react';
+import { Article } from './components/article/Article';
+import { ArticleParamsForm } from './components/article-params-form/ArticleParamsForm';
+import {
+	defaultArticleState,
+	ArticleStateType,
+	OptionType,
+} from './constants/articleProps';
+
+import './styles/index.scss';
+import styles from './styles/index.module.scss';
+
+/**
+ * компонент App корневой комопнент проекта
+ */
+
+export const App = () => {
+	const [articleState, setArticleState] =
+		useState<ArticleStateType>(defaultArticleState);
+
+	const handleChange = (key: keyof ArticleStateType, value: OptionType) => {
+		setArticleState((prev) => ({ ...prev, [key]: value }));
+	};
+
+	const handleReset = () => {
+		setArticleState(defaultArticleState);
+	};
+
+	return (
+		<main
+			className={clsx(styles.main)}
+			style={
+				{
+					'--font-family': articleState.fontFamilyOption.value,
+					'--font-size': articleState.fontSizeOption.value,
+					'--font-color': articleState.fontColor.value,
+					'--container-width': articleState.contentWidth.value,
+					'--bg-color': articleState.backgroundColor.value,
+				} as CSSProperties
+			}>
+			<ArticleParamsForm
+				articleState={articleState}
+				onChange={handleChange}
+				onReset={handleReset}
+			/>
+			<Article />
+		</main>
+	);
+};
