+++ src/index.tsx
@@ -1,54 +1,10 @@
 import { createRoot } from 'react-dom/client';
-import { StrictMode, CSSProperties, useState } from 'react';
-import clsx from 'clsx';
-
-import { Article } from './components/article/Article';
-import { ArticleParamsForm } from './components/article-params-form/ArticleParamsForm';
-import {
-	defaultArticleState,
-	ArticleStateType,
-	OptionType,
-} from './constants/articleProps';
-
+import { StrictMode } from 'react';
 import './styles/index.scss';
-import styles from './styles/index.module.scss';
+import { App } from './App';
 
 const domNode = document.getElementById('root') as HTMLDivElement;
 const root = createRoot(domNode);
-
-const App = () => {
-	const [articleState, setArticleState] =
-		useState<ArticleStateType>(defaultArticleState);
-
-	const handleChange = (key: keyof ArticleStateType, value: OptionType) => {
-		setArticleState((prev) => ({ ...prev, [key]: value }));
-	};
-
-	const handleReset = () => {
-		setArticleState(defaultArticleState);
-	};
-
-	return (
-		<main
-			className={clsx(styles.main)}
-			style={
-				{
-					'--font-family': articleState.fontFamilyOption.value,
-					'--font-size': articleState.fontSizeOption.value,
-					'--font-color': articleState.fontColor.value,
-					'--container-width': articleState.contentWidth.value,
-					'--bg-color': articleState.backgroundColor.value,
-				} as CSSProperties
-			}>
-			<ArticleParamsForm
-				articleState={articleState}
-				onChange={handleChange}
-				onReset={handleReset}
-			/>
-			<Article />
-		</main>
-	);
-};
 
 root.render(
 	<StrictMode>
